fix doubled dir name in docker run ; test before cp in make_gitian_vms.sh

+++ make_gitian_vms.sh
@@ -25,9 +25,9 @@ cd /gitian/gitian-builder
 
 cp -av /gitian/gitian-builder /data
 test -f  /data/gitian-builder/base-precise-i386  || ./bin/make-base-vm --lxc --arch i386  --suite precise 
-cp -av /gitian/gitian-builder/base-precise-i386  /data
+test -f  /data/gitian-builder/base-precise-i386  || cp -av /gitian/gitian-builder/base-precise-i386  /data
 test -f  /data/gitian-builder/base-precise-amd64 || ./bin/make-base-vm --lxc --arch amd64 --suite precise 
-cp -av /gitian/gitian-builder/base-precise-amd64  /data
+test -f  /data/gitian-builder/base-precise-i386  || cp -av /gitian/gitian-builder/base-precise-amd64  /data
 
 echo "Now you're ready to run a build"
 echo "docker run -it --rm --privileged -c $(pwd)/${GD_BUILD_COIN}-src:/gitian/${GD_BUILD_COIN} -v $(pwd)/gitian-builder:/gitian/gitian-builder  guruvan/gitian-builder"
